getting_started_with_opeanai_and_ollama/requirements.txt
langchain
ipykernel
python-dotenv
langchain-community
pypdf
bs4 

arxiv 
pymupdf
wikipedia
langchain-text-splitters



# Start from here
ipykernel
langchain
python-dotenv
langchain-openai



# 2 Build GenAI App
beautifulsoup4
langchain_community
faiss-cpu



# 3 and 4 Build GenAI using opensource models using Ollama
streamlit 



























































































































































































































































































































































































































































































































































































































































































































































































































































































































































































































































































































































































































































































































































































































































# conda activate .venv/